Remove API changes in test-infra decouple refactor
