+++ components/camel-aws/camel-aws2-iam/src/test/java/org/apache/camel/component/aws2/iam/integration/Aws2IAMBase.java
@@ -18,8 +18,8 @@ package org.apache.camel.component.aws2.iam.integration;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.component.aws2.iam.IAM2Component;
+import org.apache.camel.test.infra.aws.common.services.AWSService;
 import org.apache.camel.test.infra.aws2.clients.AWSSDKClientUtils;
-import org.apache.camel.test.infra.aws2.services.AWSService;
 import org.apache.camel.test.infra.aws2.services.AWSServiceFactory;
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.junit.jupiter.api.TestInstance;
