+++ components/camel-aws/camel-aws2-kms/src/test/java/org/apache/camel/component/aws2/kms/localstack/Aws2KmsBase.java
@@ -18,8 +18,8 @@ package org.apache.camel.component.aws2.kms.localstack;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.component.aws2.kms.KMS2Component;
+import org.apache.camel.test.infra.aws.common.services.AWSService;
 import org.apache.camel.test.infra.aws2.clients.AWSSDKClientUtils;
-import org.apache.camel.test.infra.aws2.services.AWSService;
 import org.apache.camel.test.infra.aws2.services.AWSServiceFactory;
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.junit.jupiter.api.TestInstance;
