+++ test-infra/camel-test-infra-aws-v2/pom.xml
@@ -39,6 +39,7 @@
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-test-infra-aws-common</artifactId>
             <version>${project.version}</version>
+            <type>test-jar</type>
         </dependency>
 
         <dependency>
